docstring: incorporate docstr.generate() in Docstring.get_docstring()
Move the knowledge about documentation generation in the Docstring
class. Also move the _strip() and _nest() functions that are only needed
by Docstring.get_docstring() and make the Docstring static methods.
